refactor(shape): move Shape into private _shape module
